requirements.txt
absl-py==0.10.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astroid==2.3.3
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
audioread==2.1.8
backcall==0.1.0
bidict==0.21.2
bleach==3.1.4
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==0.5.3
colorama==0.4.4
cryptography==2.8
cycler==0.10.0
dask==0.18.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
gast==0.3.3
geographiclib==1.50
geopy==2.1.0
gevent==21.1.2
gevent-websocket==0.10.1
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.22.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
httplib2==0.19.1
icc-rt==2019.0
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.1
importlib-metadata==1.6.0
intel-openmp==2019.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywebrtc==0.5.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
jupyterlab==2.0.1
jupyterlab-pygments==0.1.2
jupyterlab-server==1.1.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.3.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mkl-fft==1.0.6
mkl-random==1.0.1.1
MouseInfo==0.1.3
mysql-connector-python==8.0.22
nbclient==0.5.1
nbconvert==5.6.1
nbformat==5.0.5
nest-asyncio==1.4.2
networkx==2.4
notebook==6.0.3
numba==0.51.2
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pooch==1.2.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.13.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
pycparser==2.18
pyfakewebcam==0.1.0
PyGetWindow==0.0.9
pygit==0.1
Pygments==2.6.1
pylint==2.4.4
PyMsgBox==1.0.9
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyperclip==1.8.1
PyRect==0.1.4
pyrsistent==0.16.0
PyScreeze==0.1.26
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==4.0.1
python-socketio==5.1.0
PyTweening==1.0.3
pytz==2019.3
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-image==0.17.2
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
SoundFile==0.10.3.post1
SQLAlchemy==1.4.4
tbb==2019.0
tbb4py==2019.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.9.3
toolz==0.9.0
torch==1.6.0+cu101
torchvision==0.7.0+cu101
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
